Refactor socket stuff
