+++ mischief/util/__init__.py
@@ -12,6 +12,7 @@ from flask import Response, jsonify, current_app
 from flask_cors import CORS
 from flask_jwt_simple import JWTManager
 from flask_redis import FlaskRedis
+from flask_socketio import SocketIO, send, join_room, leave_room
 from playhouse.postgres_ext import PostgresqlExtDatabase
 from werkzeug.routing import IntegerConverter
 from os import environ
@@ -31,6 +32,8 @@ mail = MailGunner()
 
 fredis = FlaskRedis(decode_responses=True)
 
+socketio = SocketIO()
+
 
 def initialize(app):
     # handle all setup
@@ -42,6 +45,19 @@ def initialize(app):
     cors.init_app(app)
     mail.init_app(app)
     fredis.init_app(app)
+    socketio.init_app(app)
+
+    @socketio.on('join')
+    def on_join(data):
+        room = data['group_id']
+        join_room(room)
+        send('Successfully joined room: ' + room)
+
+    @socketio.on('leave')
+    def on_leave(data):
+        room = data['group_id']
+        leave_room(room)
+        send('Successfully left room: ' + room)
 
     from mischief.models.user import User
     from mischief.models.group import Group
